requirements.txt
anthropic>=0.8.0
whisper>=1.1.10
yt-dlp>=2023.3.4
torch>=2.0.0
python-dotenv>=1.0.0
mcp>=0.1.0
